// src/app/modulos/shared/domain/models/defaultOptionsHttp.ts
import { HttpHeaders, HttpParams } from "@angular/common/http";

export interface DefaultOptionsHttp {
    headers?: HttpHeaders | { [header: string]: string | string[] };
    observe?: 'body';
    params?: HttpParams | { [param:string]: string | string[] };
    reportProgress?: boolean;
    responseType?: 'json';
    withCredentials?: boolean;
}